Refactor: moved parsing fields into Part class, docx operations into own class
