+++ src/mailmerge/mailmerge.py
@@ -1,6 +1,5 @@
 import os
 import warnings
-from copy import deepcopy
 from dataclasses import dataclass
 
 # import locale
@@ -8,10 +7,10 @@ from zipfile import ZIP_DEFLATED, ZipFile
 
 from lxml import etree
 
-from .constants import CONTENT_TYPES_PARTS, MAKE_TESTS_HAPPY, NAMESPACES
-from .field import SimpleMergeField, SkipRecord
+from .constants import CONTENT_TYPES_PARTS, NAMESPACES
+from .field import SkipRecord
 from .mergedata import MergeData
-from .part import MergeDocument, MergeHeaderFooterDocument
+from .part import MergeDocument, MergeHeaderFooterDocument, Part
 from .rels import RelationsDocument
 
 
@@ -20,6 +19,79 @@ class MailMergeSettings:
     remove_empty_tables: bool
     auto_update_fields_on_open: str
     keep_fields: str
+
+
+class MailMergeDocx:
+    """
+    DOCX specific operations
+    """
+
+    def __init__(self, file):
+        self.zip = ZipFile(file)
+        self.zip_is_closed = False
+        self.parts = {}  # zi_part: ElementTree
+        self.category_part_map = {}  # category: [zi, ...]
+
+    def fill_parts(self):
+        content_types_zi = self.zip.getinfo("[Content_Types].xml")
+        content_types = etree.parse(self.zip.open(content_types_zi))
+        self.category_part_map["content_types"] = [content_types_zi]
+        self.parts[content_types_zi] = dict(part=content_types)
+        for file in content_types.findall("{%(ct)s}Override" % NAMESPACES):
+            part_type = file.attrib["ContentType" % NAMESPACES]
+            category = CONTENT_TYPES_PARTS.get(part_type)
+            if category:
+                zi, self.parts[zi] = self.__get_tree_of_file(file)
+                self.category_part_map.setdefault(category, []).append(zi)
+
+    def __get_tree_of_file(self, file):
+        fn = file.attrib["PartName" % NAMESPACES].split("/", 1)[1]
+        zi = self.zip.getinfo(fn)
+        return zi, dict(zi=zi, file=file, part=etree.parse(self.zip.open(zi)))
+
+    def get_parts(self, category_part_map=None):
+        """return all the parts based on category_part_map"""
+        if category_part_map is None:
+            category_part_map = ["main", "header_footer", "notes"]
+        elif isinstance(category_part_map, str):
+            category_part_map = [category_part_map]
+        return [self.parts[zi] for category in category_part_map for zi in self.category_part_map.get(category, [])]
+
+    def get_relations_part(self, part_zi):
+        """returns the relations document for the given part"""
+
+        rel_fn = "word/_rels/%s.rels" % os.path.basename(part_zi.filename)
+        if rel_fn in self.zip.namelist():
+            zi = self.zip.getinfo(rel_fn)
+            rel_root = etree.parse(self.zip.open(zi))
+            self.parts[zi] = dict(zi=zi, part=rel_root)
+            return rel_root
+        # else:
+        #     print(rel_fn, self.zip.namelist())
+
+    def write(self, output, new_parts):
+        for zi in self.zip.filelist:
+            if zi in self.parts:
+                xml = etree.tostring(
+                    self.parts[zi]["part"].getroot(),
+                    encoding="UTF-8",
+                    xml_declaration=True,
+                )
+                output.writestr(zi.filename, xml)
+            else:
+                output.writestr(zi.filename, self.zip.read(zi))
+
+        for new_part in new_parts:
+            xml = etree.tostring(new_part.content.getroot(), encoding="UTF-8", xml_declaration=True)
+            output.writestr(new_part.path, xml)
+            # TODO add relations
+
+    def close(self):
+        if not self.zip_is_closed:
+            try:
+                self.zip.close()
+            finally:
+                self.zip_is_closed = True
 
 
 class MailMerge(object):
@@ -53,23 +125,19 @@ class MailMerge(object):
         keep_fields : none - merge all fields even if no data, some - keep fields with no data, all - keep all fields
         """
         self.settings = MailMergeSettings(remove_empty_tables, auto_update_fields_on_open, keep_fields)
-        self.zip = ZipFile(file)
-        self.zip_is_closed = False
-        self.parts = {}  # zi_part: ElementTree
-        self.new_parts = []  # list of [(filename, part)]
-        self.categories = {}  # category: [zi, ...]
+        self.docx = MailMergeDocx(file)
         self.merge_data = MergeData(settings=self.settings)
+        self.new_parts = []  # list of [(filename, part)]
         self._has_unmerged_fields = False
 
         try:
-            self.__fill_parts()
+            self.docx.fill_parts()
 
-            for part_info in self.get_parts():
-                self.__fill_simple_fields(part_info["part"])
-                self.__fill_complex_fields(part_info["part"])
+            for part_info in self.docx.get_parts():
+                Part(self.merge_data, part_info["part"]).parse()
 
         except Exception:
-            self.zip.close()
+            self.docx.close()
             raise
 
     def __getattr__(self, name):
@@ -86,157 +154,27 @@ class MailMerge(object):
             return
         super().__setattr__(name, value)
 
-    def get_parts(self, categories=None):
-        """return all the parts based on categories"""
-        if categories is None:
-            categories = ["main", "header_footer", "notes"]
-        elif isinstance(categories, str):
-            categories = [categories]
-        return [self.parts[zi] for category in categories for zi in self.categories.get(category, [])]
-
     def get_settings(self):
         """returns the settings part"""
-        return self.parts[self.categories["settings"][0]]["part"]
+        return self.docx.parts[self.docx.category_part_map["settings"][0]]["part"]
 
     def get_content_types(self):
         """ " returns the content types part"""
-        return self.parts[self.categories["content_types"][0]]["part"]
+        return self.docx.parts[self.docx.category_part_map["content_types"][0]]["part"]
 
     def get_relations(self, part_zi):
-        """returns the"""
-        rel_fn = "word/_rels/%s.rels" % os.path.basename(part_zi.filename)
-        if rel_fn in self.zip.namelist():
-            zi = self.zip.getinfo(rel_fn)
-            rel_root = etree.parse(self.zip.open(zi))
-            self.parts[zi] = dict(zi=zi, part=rel_root)
+        """returns the relations document for the given part"""
+        rel_root = self.docx.get_relations_part(part_zi)
+        if rel_root is not None:
             relations = RelationsDocument(rel_root)
             for relation in relations.get_all():
                 self.merge_data.unique_id_manager.register_id_str(relation.attrib["Id"])
             return relations
-        # else:
-        #     print(rel_fn, self.zip.namelist())
-
-    def __fill_parts(self):
-        content_types_zi = self.zip.getinfo("[Content_Types].xml")
-        content_types = etree.parse(self.zip.open(content_types_zi))
-        self.categories["content_types"] = [content_types_zi]
-        self.parts[content_types_zi] = dict(part=content_types)
-        for file in content_types.findall("{%(ct)s}Override" % NAMESPACES):
-            part_type = file.attrib["ContentType" % NAMESPACES]
-            category = CONTENT_TYPES_PARTS.get(part_type)
-            if category:
-                zi, self.parts[zi] = self.__get_tree_of_file(file)
-                self.categories.setdefault(category, []).append(zi)
-
-    def __fill_simple_fields(self, part):
-        for fld_simple_elem in part.findall(".//{%(w)s}fldSimple" % NAMESPACES):
-            first_run_elem = deepcopy(fld_simple_elem.find("{%(w)s}r" % NAMESPACES))
-            if MAKE_TESTS_HAPPY:
-                first_run_elem.clear()
-            merge_field_obj = self.merge_data.make_data_field(
-                fld_simple_elem.getparent(),
-                instr=fld_simple_elem.get("{%(w)s}instr" % NAMESPACES),
-                field_class=SimpleMergeField,
-                all_elements=[fld_simple_elem],
-                instr_elements=[first_run_elem],
-                show_elements=[first_run_elem],
-            )
-            if merge_field_obj:
-                merge_field_obj.insert_into_tree()
-
-    def __get_next_element(self, current_element):
-        """returns the next element of a complex field"""
-        next_element = current_element.getnext()
-        current_paragraph = current_element.getparent()
-        # we search through paragraphs for the next <w:r> element
-        while next_element is None:
-            current_paragraph = current_paragraph.getnext()
-            if current_paragraph is None:
-                return None, None, None
-            next_element = current_paragraph.find("w:r", namespaces=NAMESPACES)
-
-        # print(''.join(next_element.xpath('w:instrText/text()', namespaces=NAMESPACES)))
-        field_char_subelem = next_element.find("w:fldChar", namespaces=NAMESPACES)
-        if field_char_subelem is None:
-            return next_element, None, None
-
-        return (
-            next_element,
-            field_char_subelem,
-            field_char_subelem.xpath("@w:fldCharType", namespaces=NAMESPACES)[0],
-        )
-
-    def _pull_next_merge_field(self, elements_of_type_begin, nested=False):
-        assert elements_of_type_begin
-        current_element = elements_of_type_begin.pop(0)
-        parent_element = current_element.getparent()
-        all_elements = []  # we need all the elments in case of updates
-        instr_elements = []  # the instruction part, elements that define how to get the value
-        show_elements = []  # the elements showing the current value
-
-        current_element_list = instr_elements
-        all_elements.append(current_element)
-
-        # good_elements = []
-        # ignore_elements = [current_element]
-        # current_element_list = good_elements
-        field_char_type = None
-
-        # print('>>>>>>>')
-        while field_char_type != "end":
-            # find next sibling
-            next_element, field_char_subelem, field_char_type = self.__get_next_element(current_element)
-
-            if next_element is None:
-                instr_text = self.merge_data.get_instr_text(instr_elements, recursive=True)
-                raise ValueError("begin without end near:" + instr_text)
-
-            if field_char_type == "begin":
-                # nested elements
-                assert elements_of_type_begin[0] is next_element
-                merge_field_sub_obj, next_element = self._pull_next_merge_field(elements_of_type_begin, nested=True)
-                if merge_field_sub_obj:
-                    next_element = merge_field_sub_obj.insert_into_tree()
-                # print("current list is ignore", current_element_list is ignore_elements)
-                # print("<<<<< #####", etree.tostring(next_element))
-            elif field_char_type == "separate":
-                current_element_list = show_elements
-            elif next_element.tag == "MergeField":
-                # we have a nested simple Field - mark it as nested
-                self.merge_data.mark_field_as_nested(next_element.get("merge_key"))
-
-            if field_char_type not in ["end", "separate"]:
-                current_element_list.append(next_element)
-            all_elements.append(next_element)
-            current_element = next_element
-
-        # print('<<<<<<<', len(good_elements), len(ignore_elements))
-        merge_obj = self.merge_data.make_data_field(
-            parent_element,
-            nested=nested,
-            all_elements=all_elements,
-            instr_elements=instr_elements,
-            show_elements=show_elements,
-        )
-        return merge_obj, current_element
-
-    def __fill_complex_fields(self, part):
-        """finds all begin fields and then builds the MergeField objects and inserts the replacement
-        Elements in the tree"""
-        # will find all "runs" containing an element of fldChar type=begin
-        elements_of_type_begin = list(
-            part.findall('.//{%(w)s}r/{%(w)s}fldChar[@{%(w)s}fldCharType="begin"]/..' % NAMESPACES)
-        )
-        while elements_of_type_begin:
-            merge_field_obj, _ = self._pull_next_merge_field(elements_of_type_begin)
-            if merge_field_obj:
-                # print(merge_field_obj.instr)
-                merge_field_obj.insert_into_tree()
 
     def __fix_settings(self):
-        settings = self.get_settings()
-        if settings:
-            settings_root = settings.getroot()
+        settings_part = self.get_settings()
+        if settings_part:
+            settings_root = settings_part.getroot()
             if not self._has_unmerged_fields:
                 mail_merge = settings_root.find("{%(w)s}mailMerge" % NAMESPACES)
                 if mail_merge is not None:
@@ -254,11 +192,6 @@ class MailMerge(object):
                         settings_root, "{%(w)s}updateFields" % NAMESPACES, attrib=None, nsmap=None
                     )
                 update_fields_elem.set("{%(w)s}val" % NAMESPACES, "true")
-
-    def __get_tree_of_file(self, file):
-        fn = file.attrib["PartName" % NAMESPACES].split("/", 1)[1]
-        zi = self.zip.getinfo(fn)
-        return zi, dict(zi=zi, file=file, part=etree.parse(self.zip.open(zi)))
 
     def write(self, file, empty_value=""):
         self._has_unmerged_fields = bool(self.get_merge_fields())
@@ -282,21 +215,7 @@ class MailMerge(object):
             content_types.append(new_part.part_content_type)
 
         with ZipFile(file, "w", ZIP_DEFLATED) as output:
-            for zi in self.zip.filelist:
-                if zi in self.parts:
-                    xml = etree.tostring(
-                        self.parts[zi]["part"].getroot(),
-                        encoding="UTF-8",
-                        xml_declaration=True,
-                    )
-                    output.writestr(zi.filename, xml)
-                else:
-                    output.writestr(zi.filename, self.zip.read(zi))
-
-            for new_part in self.new_parts:
-                xml = etree.tostring(new_part.content.getroot(), encoding="UTF-8", xml_declaration=True)
-                output.writestr(new_part.path, xml)
-                # TODO add relations
+            self.docx.write(output, self.new_parts)
 
     def get_merge_fields(self):
         """ " get the fields from the document"""
@@ -304,7 +223,7 @@ class MailMerge(object):
 
     def _get_merge_fields(self, parts=None):
         if not parts:
-            parts = self.get_parts()
+            parts = self.docx.get_parts()
 
         fields = set()
         for part in parts:
@@ -336,7 +255,7 @@ class MailMerge(object):
 
         # prepare the side documents, like headers, footers, etc
         rel_docs = []
-        for part_info in self.get_parts(["header_footer"]):
+        for part_info in self.docx.get_parts(["header_footer"]):
             relations = self.get_relations(part_info["zi"])
             merge_header_footer_doc = MergeHeaderFooterDocument(part_info, relations, separator)
             rel_docs.append(merge_header_footer_doc)
@@ -348,7 +267,7 @@ class MailMerge(object):
         # a new break or a new section break.
 
         # GET ROOT - WORK WITH DOCUMENT
-        for part_info in self.get_parts(["main"]):
+        for part_info in self.docx.get_parts(["main"]):
             root = part_info["part"].getroot()
             relations = self.get_relations(part_info["zi"])
 
@@ -396,7 +315,7 @@ class MailMerge(object):
         self._merge(replacements)
 
     def _merge(self, replacements):
-        for part_info in self.get_parts():
+        for part_info in self.docx.get_parts():
             self.merge_data.replace(part_info["part"], replacements)
 
         for new_part in self.new_parts:
@@ -405,18 +324,19 @@ class MailMerge(object):
     def merge_rows(self, anchor, rows):
         """anchor is one of the fields in the table"""
 
-        for part_info in self.get_parts():
+        for part_info in self.docx.get_parts():
             self.merge_data.replace_table_rows(part_info["part"], anchor, rows)
 
     def __enter__(self):
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
-        self.close()
+        self.docx.close()
 
     def close(self):
-        if not self.zip_is_closed:
-            try:
-                self.zip.close()
-            finally:
-                self.zip_is_closed = True
+        self.docx.close()
+        warnings.warn(
+            "close() has been deprecated. Use the *with* statement.",
+            category=DeprecationWarning,
+            stacklevel=2,
+        )
